Create BODY.py
A Pyrthon programme that uses opencv and Mediapipe to track the movements of your body with a webcam. Clicking on the python file will automatically install the required packages and then laaunch the programme. It may only install the required packages and won't launch the programme. Close the window and re-open the python file, it should launch.
Git stats